Updated alternative routes for Auto
