+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/FreightFrenzy_2021/competition/Mecanum_Auto_BlueBarrier.java
@@ -125,13 +125,8 @@ public class Mecanum_Auto_BlueBarrier extends LinearOpMode {
         telemetry.addData(">", "Press Play to start op mode");
         telemetry.update();
 
-        //Traj
+        //TRAJ
         SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
-        Pose2d startPose = new Pose2d(6.5, 62.125, toRadians(90));
-        drive.setPoseEstimate(startPose);
-        Trajectory myTrajectory1 = drive.trajectoryBuilder(startPose,true)
-                .splineToConstantHeading(new Vector2d(-11.875, 43), toRadians(-90))
-                .build();
 
         waitForStart();
         if(opModeIsActive()) {
@@ -159,19 +154,43 @@ public class Mecanum_Auto_BlueBarrier extends LinearOpMode {
                     }
                 }
             }
+            Pose2d startPose = fieldConstant.BLUE_BARRIER_STARTING_POSE;
+            drive.setPoseEstimate(startPose);
             if (center < 0) {
                 visionResult = "LEFT";
             } else if (center < 321.85) {
                 visionResult = "MIDDLE";
+                Trajectory plateTraj0 = drive.trajectoryBuilder(startPose,true)
+                        .strafeRight(5)
+                        .build();
+                drive.followTrajectory(plateTraj0);
+                startPose = plateTraj0.end();
             } else {
                 visionResult = "RIGHT";
+                Trajectory plateTraj0 = drive.trajectoryBuilder(startPose,true)
+                        .strafeRight(5)
+                        .build();
+                drive.followTrajectory(plateTraj0);
+                startPose = plateTraj0.end();
             }
             telemetry.addLine(visionResult);
             telemetry.update();
 
+            //MOVE MARKER OUT
+            Trajectory plateTraj1 = drive.trajectoryBuilder(startPose,true)
+                    .back(45)
+                    .build();
+            drive.followTrajectory(plateTraj1);
+            Trajectory plateTraj2 = drive.trajectoryBuilder(plateTraj1.end())
+                    .forward(2)
+                    .build();
+            drive.followTrajectory(plateTraj2);
+
             //MOTION TO PLATE
-            drive.followTrajectory(myTrajectory1);
-            sleep(600);
+            Trajectory plateTraj3 = drive.trajectoryBuilder(plateTraj2.end())
+                    .lineToLinearHeading(new Pose2d(7.875, 23.75, toRadians(0)))
+                    .build();
+            drive.followTrajectory(plateTraj3);
 
             //ROTATE
             Rotate.setPosition(1.0);
@@ -194,10 +213,11 @@ public class Mecanum_Auto_BlueBarrier extends LinearOpMode {
             sleep(600);
 
             //CLOSER TO PLATE
-            Trajectory myTrajectory2 = drive.trajectoryBuilder(myTrajectory1.end())
-                    .back(2)
+            Trajectory closerTraj = drive.trajectoryBuilder(plateTraj3.end())
+                    .back(1.5)
                     .build();
-            drive.followTrajectory(myTrajectory2);
+            drive.followTrajectory(closerTraj);
+            drive.setPoseEstimate(closerTraj.end());
 
             //DUMP AND SLIDE DOWN
             Push.setPosition(0.0);
@@ -211,382 +231,33 @@ public class Mecanum_Auto_BlueBarrier extends LinearOpMode {
             Slide.setPower(0.8);
 
             //BACK TO WALL
-            Trajectory myTrajectory3 = drive.trajectoryBuilder(myTrajectory2.end())
-                    .splineToLinearHeading(new Pose2d(6.5, 65), toRadians(-180))
+            if(visionResult == "RIGHT"){
+                Trajectory wallTrajR1 = drive.trajectoryBuilder(closerTraj.end())
+                        .lineToLinearHeading(new Pose2d(14.8, 23.75, toRadians(90)))
+                        .build();
+                drive.followTrajectory(wallTrajR1);
+                Trajectory wallTrajR2 = drive.trajectoryBuilder(wallTrajR1.end())
+                        .forward(20)
+                        .build();
+                drive.followTrajectory(wallTrajR2);
+                drive.setPoseEstimate(wallTrajR2.end());
+            }
+            Trajectory wallTraj1 = drive.trajectoryBuilder(drive.getPoseEstimate())
+                    .lineToLinearHeading(new Pose2d(6.5, 65, toRadians(0)))
                     .build();
-            drive.followTrajectory(myTrajectory3);
+            drive.followTrajectory(wallTraj1);
             sleep(300);
 
-            Trajectory myTrajectory4 = drive.trajectoryBuilder(myTrajectory3.end())
-                    .forward(35)
+            Trajectory wallTraj2 = drive.trajectoryBuilder(wallTraj1.end())
+                    .forward(36)
                     .build();
-            drive.followTrajectory(myTrajectory4);
+            drive.followTrajectory(wallTraj2);
 
+
+            PoseStorage.currentPose = wallTraj1.end();
+            PoseStorage.state = driveMethod.poseState.BLUE;
         }
 
-    }
-    void stopMotion() {
-        LF.setPower(0);
-        RF.setPower(0);
-        LB.setPower(0);
-        RB.setPower(0);
-    }
-
-    void stopMotion(long timeInterval) {
-        stopMotion();
-        sleep(timeInterval);
-    }
-
-    double normalizeAngle(double angle) {
-        double tempDeg = angle % 360;
-        if (tempDeg >= 180) {
-            tempDeg -= 360;
-        } else if (tempDeg < -180) {
-            tempDeg += 360;
-        }
-        return tempDeg;
-    }
-
-    double aquireHeading() {
-        Orientation angles = imu.getAngularOrientation(AxesReference.INTRINSIC, AxesOrder.ZYX, AngleUnit.DEGREES);
-        double heading = angles.firstAngle;
-        double tempHead = normalizeAngle(heading);
-        telemetry.addData("Heading", tempHead);
-        telemetry.update();
-        sleep(20);
-        return tempHead;
-    }
-
-    void displayEncoderValue() {
-        try {
-            double LFDis = LF.getCurrentPosition() / 537.7 * 1.89 * 2 * 3.1415926;
-            double RFDis = RF.getCurrentPosition() / 537.7 * 1.89 * 2 * 3.1415926;
-            double LBDis = LB.getCurrentPosition() / 537.7 * 1.89 * 2 * 3.1415926;
-            double RBDis = RB.getCurrentPosition() / 537.7 * 1.89 * 2 * 3.1415926;
-            telemetry.addData("LF Encoder Value: ", LFDis);
-            telemetry.addData("RF Encoder Value: ", RFDis);
-            telemetry.addData("LB Encoder Value: ", LBDis);
-            telemetry.addData("RB Encoder Value: ", RBDis);
-            telemetry.addData("X Value: ", (LFDis / Math.sqrt(2.0) + RBDis / Math.sqrt(2.0) - LBDis / Math.sqrt(2.0) - RFDis / Math.sqrt(2.0)));
-            telemetry.addData("Y Value: ", (LFDis / Math.sqrt(2.0) + RFDis / Math.sqrt(2.0) + LBDis / Math.sqrt(2.0) + RBDis / Math.sqrt(2.0)));
-            telemetry.addData("Average Encoder Value: ", (LFDis + LBDis + RFDis + RBDis) / 4.0);
-            telemetry.update();
-        } catch (Exception e) {
-            telemetry.addLine("Unable to find encoder value");
-            telemetry.update();
-        }
-    }
-
-    void driveStraight(boolean isForward, double margin, double power, double timeInterval) {
-        ElapsedTime driveTime = new ElapsedTime(ElapsedTime.Resolution.MILLISECONDS);
-        final double currentAngle = aquireHeading();
-        int straightFactor = -1;
-        if (isForward) {
-            straightFactor = 1;
-        }
-        double targetAngle = currentAngle;
-        double LF_power;
-        double LB_power;
-        double RF_power;
-        double RB_power;
-        while (driveTime.milliseconds() < timeInterval) {
-            double tempAngle = aquireHeading();
-            LF_power = straightFactor * power;
-            LB_power = straightFactor * power;
-            RF_power = straightFactor * power;
-            RB_power = straightFactor * power;
-            if (tempAngle < normalizeAngle(targetAngle - 1 * margin)) {
-                RF_power += 0.1;
-                RB_power += 0.1;
-                LF_power -= 0.1;
-                LB_power -= 0.1;
-            } else if (tempAngle > normalizeAngle(targetAngle + (margin))) {
-                RF_power -= 0.1;
-                RB_power -= 0.1;
-                LF_power += 0.1;
-                LB_power += 0.1;
-            }
-            RF_power = Range.clip(RF_power, -1, 1);
-            RB_power = Range.clip(RB_power, -1, 1);
-            LF_power = Range.clip(LF_power, -1, 1);
-            LB_power = Range.clip(LB_power, -1, 1);
-            LF.setPower(LF_power);
-            RF.setPower(RF_power);
-            LB.setPower(LB_power);
-            RB.setPower(RB_power);
-            telemetry.addData("RF_power", RF_power);
-            telemetry.addData("RB_power", RB_power);
-            telemetry.addData("LF_power", LF_power);
-            telemetry.addData("LB_power", LB_power);
-            telemetry.update();
-            displayEncoderValue();
-        }
-        stopMotion();
-    }
-
-//    void encoderDrive(double LF_D, double RF_D, double LB_D, double RB_D){
-//        LF_D/
-//    }
-
-    void drivePerpendicularly(boolean isLeft, double margin, double power, double timeInterval) {
-        ElapsedTime driveTime = new ElapsedTime(ElapsedTime.Resolution.MILLISECONDS);
-        final double currentAngle = aquireHeading();
-        int perpendicularFactor = -1;
-        if (isLeft) {
-            perpendicularFactor = 1;
-        }
-        double targetAngle = normalizeAngle(currentAngle + 90 * perpendicularFactor);
-        double LF_power;
-        double LB_power;
-        double RF_power;
-        double RB_power;
-        while (driveTime.milliseconds() < timeInterval) {
-            double tempAngle = aquireHeading();
-            LF_power = -1 * perpendicularFactor * power;
-            LB_power = perpendicularFactor * power;
-            RF_power = perpendicularFactor * power;
-            RB_power = -1 * perpendicularFactor * power;
-            if (tempAngle < normalizeAngle(targetAngle - 1 * margin)) {
-                RF_power += perpendicularFactor * 0.1;
-                RB_power -= perpendicularFactor * 0.1;
-                LF_power += perpendicularFactor * 0.1;
-                LB_power -= perpendicularFactor * 0.1;
-            } else if (tempAngle > normalizeAngle(targetAngle + (margin))) {
-                RF_power -= perpendicularFactor * 0.1;
-                RB_power += perpendicularFactor * 0.1;
-                LF_power -= perpendicularFactor * 0.1;
-                LB_power += perpendicularFactor * 0.1;
-            }
-            RF_power = Range.clip(RF_power, -1, 1);
-            RB_power = Range.clip(RB_power, -1, 1);
-            LF_power = Range.clip(LF_power, -1, 1);
-            LB_power = Range.clip(LB_power, -1, 1);
-            LF.setPower(LF_power);
-            RF.setPower(RF_power);
-            LB.setPower(LB_power);
-            RB.setPower(RB_power);
-            telemetry.addData("RF_power", RF_power);
-            telemetry.addData("RB_power", RB_power);
-            telemetry.addData("LF_power", LF_power);
-            telemetry.addData("LB_power", LB_power);
-            telemetry.update();
-        }
-        stopMotion();
-    }
-
-    void driveLeftTurn(boolean isForward, double margin, double power, double timeInterval) {
-        ElapsedTime driveTime = new ElapsedTime(ElapsedTime.Resolution.MILLISECONDS);
-        final double currentAngle = aquireHeading();
-        int straightFactor = -1;
-        if (isForward) {
-            straightFactor = 1;
-        }
-        double LF_power;
-        double LB_power;
-        double RF_power;
-        double RB_power;
-        while (driveTime.milliseconds() < timeInterval) {
-            double tempAngle = aquireHeading();
-            LF_power = straightFactor * power;
-            LB_power = straightFactor * power;
-            RF_power = straightFactor * power;
-            RB_power = straightFactor * power;
-            RF_power = Range.clip(RF_power, -1, 1);
-            RB_power = Range.clip(RB_power, -1, 1);
-            LF_power = Range.clip(LF_power, -1, 1);
-            LB_power = Range.clip(LB_power, -1, 1);
-            LF.setPower(LF_power);
-            RF.setPower(-RF_power);
-            LB.setPower(LB_power);
-            RB.setPower(-RB_power);
-            telemetry.addData("RF_power", -RF_power);
-            telemetry.addData("RB_power", -RB_power);
-            telemetry.addData("LF_power", LF_power);
-            telemetry.addData("LB_power", LB_power);
-
-            telemetry.update();
-            displayEncoderValue();
-        }
-        stopMotion();
-    }
-
-    void driveRightTurn(boolean isForward, double margin, double power, double timeInterval) {
-        ElapsedTime driveTime = new ElapsedTime(ElapsedTime.Resolution.MILLISECONDS);
-        final double currentAngle = aquireHeading();
-        int straightFactor = -1;
-        if (isForward) {
-            straightFactor = 1;
-        }
-        double targetAngle = currentAngle;
-        double LF_power;
-        double LB_power;
-        double RF_power;
-        double RB_power;
-        while (driveTime.milliseconds() < timeInterval) {
-
-            LF_power = straightFactor * power;
-            LB_power = straightFactor * power;
-            RF_power = straightFactor * power;
-            RB_power = straightFactor * power;
-
-            RF_power = Range.clip(RF_power, -1, 1);
-            RB_power = Range.clip(RB_power, -1, 1);
-            LF_power = Range.clip(LF_power, -1, 1);
-            LB_power = Range.clip(LB_power, -1, 1);
-            LF.setPower(-LF_power);
-            RF.setPower(RF_power);
-            LB.setPower(-LB_power);
-            RB.setPower(RB_power);
-            telemetry.addData("RF_power", RF_power);
-            telemetry.addData("RB_power", RB_power);
-            telemetry.addData("LF_power", -LF_power);
-            telemetry.addData("LB_power", -LB_power);
-
-            telemetry.update();
-            displayEncoderValue();
-        }
-        stopMotion();
-    }
-    void driveStrafeLeft(boolean isForward, double margin, double power, double timeInterval) {
-        ElapsedTime driveTime = new ElapsedTime(ElapsedTime.Resolution.MILLISECONDS);
-        final double currentAngle = aquireHeading();
-        int straightFactor = -1;
-        if (isForward) {
-            straightFactor = 1;
-        }
-        double targetAngle = currentAngle;
-        double LF_power;
-        double LB_power;
-        double RF_power;
-        double RB_power;
-        while (driveTime.milliseconds() < timeInterval) {
-
-            LF_power = straightFactor * power;
-            LB_power = straightFactor * power;
-            RF_power = straightFactor * power;
-            RB_power = straightFactor * power;
-
-            RF_power = Range.clip(RF_power, -1, 1);
-            RB_power = Range.clip(RB_power, -1, 1);
-            LF_power = Range.clip(LF_power, -1, 1);
-            LB_power = Range.clip(LB_power, -1, 1);
-            LF.setPower(0);
-            RF.setPower(RF_power);
-            LB.setPower(-LB_power);
-            RB.setPower(0);
-            telemetry.addData("RF_power", RF_power);
-            telemetry.addData("RB_power", RB_power);
-            telemetry.addData("LF_power", -LF_power);
-            telemetry.addData("LB_power", -LB_power);
-
-            telemetry.update();
-            displayEncoderValue();
-        }
-        stopMotion();
-    }
-
-    //make a turn that based on the current heading in a certain direction and angle
-    void rotateAtAngle(boolean isClockwise, double degree, double margin, double power) {
-        int angleFactor = -1;
-        if (!isClockwise) {
-            angleFactor = 1;
-        }
-        final double currentAngle = aquireHeading();
-        double targetAngle = normalizeAngle(currentAngle + degree * angleFactor);
-        rotateToAngle(targetAngle, margin, power);
-        stopMotion();
-    }
-
-    //make a turn TO a certain angle
-    void rotateToAngle(double targetAngle, double margin, double power) {
-        int angleFactor = 0;
-        final double currentAngle = aquireHeading();
-        if (currentAngle - targetAngle > 0) {
-            if (currentAngle - targetAngle < 180) {
-                //cw
-                angleFactor = -1;
-            } else {
-                //ccw
-                angleFactor = 1;
-            }
-        } else {
-            if (targetAngle - currentAngle < 180) {
-                //ccw
-                angleFactor = 1;
-            } else {
-                //cw
-                angleFactor = -1;
-            }
-        }
-        double LF_power;
-        double LB_power;
-        double RF_power;
-        double RB_power;
-        double tempAngle = currentAngle;
-        while (!((tempAngle < targetAngle + margin) && (tempAngle > targetAngle - margin))) {
-            tempAngle = aquireHeading();
-            RF_power = angleFactor * power;
-            RB_power = angleFactor * power;
-            LF_power = -1 * angleFactor * power;
-            LB_power = -1 * angleFactor * power;
-            RF_power = Range.clip(RF_power, -1, 1);
-            RB_power = Range.clip(RB_power, -1, 1);
-            LF_power = Range.clip(LF_power, -1, 1);
-            LB_power = Range.clip(LB_power, -1, 1);
-            LF.setPower(LF_power);
-            RF.setPower(RF_power);
-            LB.setPower(LB_power);
-            RB.setPower(RB_power);
-            telemetry.addData("RF_power", RF_power);
-            telemetry.addData("RB_power", RB_power);
-            telemetry.addData("LF_power", LF_power);
-            telemetry.addData("LB_power", LB_power);
-            telemetry.update();
-        }
-        stopMotion();
-    }
-
-    boolean slideMotion(int targetPosition, double power) {
-        Slide.setTargetPosition(targetPosition);
-        Slide.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-        Slide.setPower(power);
-        telemetry.addLine("Slide Current: " + Slide.getCurrentPosition());
-        telemetry.addLine("Slide Target: " + Slide.getTargetPosition());
-        telemetry.update();
-        return Slide.isBusy();
-    }
-
-    boolean rotateMotion(boolean moveOut) {
-        if(moveOut)
-            Rotate.setPosition(1.0);
-        else
-            Rotate.setPosition(0.03);
-        telemetry.addLine("Rotate: " + Rotate.getPosition());
-        telemetry.update();
-        sleep(300);
-        return (Rotate.getPosition() == 1.0 || Rotate.getPosition() == 0.03);
-    }
-
-    boolean pushMotion(boolean moveOut) {
-        if(moveOut)
-            Push.setPosition(0.4);
-        else
-            Push.setPosition(0.0);
-        telemetry.addLine("Push: " + Push.getPosition());
-        telemetry.update();
-        sleep(300);
-        return (Push.getPosition() == 0.0 || Push.getPosition() == 0.4);
-    }
-
-
-
-    boolean spinMotion(double power, double timeInterval) {
-        ElapsedTime spinTime = new ElapsedTime(ElapsedTime.Resolution.MILLISECONDS);
-        while (spinTime.milliseconds() <= timeInterval) {
-            Spin.setPower(power);
-        }
-        Spin.setPower(0);
-        return Spin.isBusy();
     }
 
     //Vision
