+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/FreightFrenzy_2021/competition/Mecanum_Auto_BlueDuck_Warehouse.java
@@ -6,10 +6,13 @@ import com.acmerobotics.roadrunner.trajectory.Trajectory;
 import com.qualcomm.hardware.bosch.BNO055IMU;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.Servo;
-import com.qualcomm.robotcore.util.ElapsedTime;
-import com.qualcomm.robotcore.util.Range;
+        import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+        import com.qualcomm.robotcore.hardware.CRServo;
+        import com.qualcomm.robotcore.hardware.DcMotor;
+        import com.qualcomm.robotcore.hardware.DcMotorSimple;
+        import com.qualcomm.robotcore.hardware.Servo;
+        import com.qualcomm.robotcore.util.ElapsedTime;
+        import com.qualcomm.robotcore.util.Range;
 
 import org.firstinspires.ftc.robotcore.external.ClassFactory;
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
@@ -28,8 +31,8 @@ import java.util.List;
 
 import static java.lang.Math.toRadians;
 
-@Autonomous(name = "RED BARRIER", group = "Competition")
-public class Mecanum_Auto_RedBarrier extends LinearOpMode {
+@Autonomous(name = "BLUE DUCK - WAREHOUSE", group = "Competition")
+public class Mecanum_Auto_BlueDuck_Warehouse extends LinearOpMode {
 
     private DcMotor LF = null;
     private DcMotor RF = null;
@@ -40,7 +43,6 @@ public class Mecanum_Auto_RedBarrier extends LinearOpMode {
     private DcMotor Slide = null;
     private Servo Rotate = null;
     private Servo Push = null;
-    private ArrayList<Double[]> speedList = new ArrayList<Double[]>();
     private ElapsedTime runtime = new ElapsedTime();
 
     BNO055IMU imu;
@@ -56,10 +58,6 @@ public class Mecanum_Auto_RedBarrier extends LinearOpMode {
             Robot4100Common.VUFORIA_LICENSE;
     private VuforiaLocalizer vuforia;
     private TFObjectDetector tfod;
-
-    double rotate = 0;
-    double speed = 0.5;
-    boolean reverse = false;
 
     @Override
     public void runOpMode() {
@@ -102,7 +100,7 @@ public class Mecanum_Auto_RedBarrier extends LinearOpMode {
         //initialize position
         Push.setPosition(0.4);
         Slide.setPower(0.15);
-        sleep(200);
+        sleep(100);
         Slide.setPower(0.0);
         Slide.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         Rotate.setPosition(0.03);
@@ -124,6 +122,11 @@ public class Mecanum_Auto_RedBarrier extends LinearOpMode {
 
         //Traj
         SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
+        Pose2d startPose = new Pose2d(-41, 62.125, toRadians(90));
+        drive.setPoseEstimate(startPose);
+        Trajectory myTrajectory1 = drive.trajectoryBuilder(startPose,true)
+                .splineToConstantHeading(new Vector2d(-11.875, 43), toRadians(-90))
+                .build();
 
         waitForStart();
         if(opModeIsActive()) {
@@ -151,43 +154,44 @@ public class Mecanum_Auto_RedBarrier extends LinearOpMode {
                     }
                 }
             }
-            Pose2d startPose = fieldConstant.RED_BARRIER_STARTING_POSE;
-            drive.setPoseEstimate(startPose);
             if (center < 0) {
-                visionResult = "RIGHT";
-            } else if (center < 420.725) {
                 visionResult = "LEFT";
-                Trajectory plateTraj0 = drive.trajectoryBuilder(startPose,true)
-                        .strafeLeft(5)
-                        .build();
-                drive.followTrajectory(plateTraj0);
-                startPose = plateTraj0.end();
-            } else {
+            } else if (center < 321.85) {
                 visionResult = "MIDDLE";
-                Trajectory plateTraj0 = drive.trajectoryBuilder(startPose,true)
-                        .strafeLeft(5)
-                        .build();
-                drive.followTrajectory(plateTraj0);
-                startPose = plateTraj0.end();
+            } else {
+                visionResult = "RIGHT";
             }
             telemetry.addLine(visionResult);
             telemetry.update();
 
-            //MOVE MARKER OUT
-            Trajectory plateTraj1 = drive.trajectoryBuilder(startPose,true)
-                    .back(45)
+            //MOTION TO DUCK
+            Trajectory duckTraj = drive.trajectoryBuilder(startPose,true)
+                    .splineToLinearHeading(new Pose2d(-64, 54, toRadians(90)), toRadians(0))
                     .build();
-            drive.followTrajectory(plateTraj1);
-            Trajectory plateTraj2 = drive.trajectoryBuilder(plateTraj1.end())
-                    .forward(2)
-                    .build();
-            drive.followTrajectory(plateTraj2);
+            drive.followTrajectory(duckTraj);
+            sleep(300);
+
+            Pose2d wall = new Pose2d(-63.88, 53.25, Math.toRadians(90));
+            drive.setPoseEstimate(wall);
+
+            //ROTATE DUCK
+            drive.setMotorPowers(0.1, 0.1,0.1,0.1);
+            sleep(500);
+            drive.setMotorPowers(0, 0,0,0);
+            Spin.setPower(0.5);
+            sleep(3500);
+            Spin.setPower(0.0);
 
             //MOTION TO PLATE
-            Trajectory plateTraj3 = drive.trajectoryBuilder(plateTraj2.end())
-                    .lineToLinearHeading(new Pose2d(7.875, -23.75, toRadians(0)))
+            Trajectory wallTraj = drive.trajectoryBuilder(duckTraj.end(), true)
+                    .splineToLinearHeading(new Pose2d(-59, 23.75, toRadians(180)), toRadians(-90))
                     .build();
-            drive.followTrajectory(plateTraj3);
+            drive.followTrajectory(wallTraj);
+
+            Trajectory plateTraj = drive.trajectoryBuilder(wallTraj.end())
+                    .back(26.375)
+                    .build();
+            drive.followTrajectory(plateTraj);
 
             //ROTATE
             Rotate.setPosition(1.0);
@@ -210,11 +214,10 @@ public class Mecanum_Auto_RedBarrier extends LinearOpMode {
             sleep(600);
 
             //CLOSER TO PLATE
-            Trajectory closerTraj = drive.trajectoryBuilder(plateTraj3.end())
-                    .back(1.5)
+            Trajectory closeTraj = drive.trajectoryBuilder(plateTraj.end())
+                    .back(1)
                     .build();
-            drive.followTrajectory(closerTraj);
-            drive.setPoseEstimate(closerTraj.end());
+            drive.followTrajectory(closeTraj);
 
             //DUMP AND SLIDE DOWN
             Push.setPosition(0.0);
@@ -227,36 +230,30 @@ public class Mecanum_Auto_RedBarrier extends LinearOpMode {
             Slide.setMode(DcMotor.RunMode.RUN_TO_POSITION);
             Slide.setPower(0.8);
 
-            //BACK TO WALL
-            if(visionResult == "LEFT"){
-                Trajectory wallTrajR1 = drive.trajectoryBuilder(closerTraj.end())
-                        .lineToLinearHeading(new Pose2d(14.8, -23.75, toRadians(-90)))
-                        .build();
-                drive.followTrajectory(wallTrajR1);
-                Trajectory wallTrajR2 = drive.trajectoryBuilder(wallTrajR1.end())
-                        .forward(20)
-                        .build();
-                drive.followTrajectory(wallTrajR2);
-                drive.setPoseEstimate(wallTrajR2.end());
-            }
-            Trajectory wallTraj1 = drive.trajectoryBuilder(drive.getPoseEstimate())
-                    .lineToLinearHeading(new Pose2d(6.5, -65, toRadians(0)))
+            //BACK TO WAREHOUSE & PARK
+            Trajectory parkTraj1 = drive.trajectoryBuilder(closeTraj.end())
+                    .strafeRight(30)
                     .build();
-            drive.followTrajectory(wallTraj1);
-            sleep(300);
-
-            Trajectory wallTraj2 = drive.trajectoryBuilder(wallTraj1.end())
-                    .forward(36)
+            drive.followTrajectory(parkTraj1);
+            Trajectory parkTraj2 = drive.trajectoryBuilder(parkTraj1.end())
+                    .strafeLeft(5)
                     .build();
-            drive.followTrajectory(wallTraj2);
+            drive.followTrajectory(parkTraj2);
+            Trajectory parkTraj3 = drive.trajectoryBuilder(parkTraj2.end())
+                    .lineToLinearHeading(new Pose2d(-31.625, 45.75, toRadians(0)))
+                    .build();
+            drive.followTrajectory(parkTraj3);
+            Trajectory parkTraj4 = drive.trajectoryBuilder(parkTraj3.end())
+                    .forward(75)
+                    .build();
+            drive.followTrajectory(parkTraj4);
+            sleep(500);
 
-
-            PoseStorage.currentPose = wallTraj1.end();
-            PoseStorage.state = driveMethod.poseState.RED;
+            PoseStorage.currentPose = parkTraj4.end();
+            PoseStorage.state = driveMethod.poseState.BLUE;
         }
 
     }
-
     //Vision
     private void initVuforia() {
         /*
@@ -286,4 +283,7 @@ public class Mecanum_Auto_RedBarrier extends LinearOpMode {
         tfod = ClassFactory.getInstance().createTFObjectDetector(tfodParameters, vuforia);
         tfod.loadModelFromAsset(TFOD_MODEL_ASSET, LABELS);
     }
+
+
+
 }
