+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/FreightFrenzy_2021/competition/Mecanum_TeleOp_Final.java
@@ -1,5 +1,8 @@
 package org.firstinspires.ftc.teamcode.FreightFrenzy_2021.competition;
 
+import com.acmerobotics.roadrunner.geometry.Pose2d;
+import com.acmerobotics.roadrunner.geometry.Vector2d;
+import com.acmerobotics.roadrunner.trajectory.Trajectory;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.CRServo;
@@ -9,7 +12,11 @@ import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
 import com.qualcomm.robotcore.util.Range;
 
+import org.firstinspires.ftc.teamcode.roadrunner.drive.SampleMecanumDrive;
+
 import java.util.ArrayList;
+
+import static java.lang.Math.toRadians;
 
 @TeleOp(name = "TELEOP FINAL", group = "Competition")
 public class Mecanum_TeleOp_Final extends LinearOpMode {
@@ -27,7 +34,7 @@ public class Mecanum_TeleOp_Final extends LinearOpMode {
     private ElapsedTime runtime = new ElapsedTime();
 
     double rotate = 0;
-    double speed = 0.8;
+    double speed = 0.7;
     boolean reverse = false;
 
     @Override
@@ -68,6 +75,10 @@ public class Mecanum_TeleOp_Final extends LinearOpMode {
         double spinPower = 0;
         int initialHeight = Slide.getCurrentPosition();
 
+        //Read Position From Auto
+        SampleMecanumDrive chassis = new SampleMecanumDrive(hardwareMap);
+        chassis.setPoseEstimate(PoseStorage.currentPose);
+
         telemetry.addData("Status", "Initialized");
         telemetry.update();
 
@@ -84,7 +95,9 @@ public class Mecanum_TeleOp_Final extends LinearOpMode {
         boolean releasedA2 = true;
         boolean releasedB1 = true;
         boolean releasedB2 = true;
+        boolean releasedY1 = true;
         boolean releasedY2 = true;
+        boolean releasedX2 = true;
         boolean releasedDD1 = true;
         boolean releasedDD2 = true;
         boolean releasedDL2 = true;
@@ -99,11 +112,17 @@ public class Mecanum_TeleOp_Final extends LinearOpMode {
 
         while (opModeIsActive()) {
             runtime.reset();
+            chassis.update();
+
+            // Retrieve your pose
+            Pose2d myPose = chassis.getPoseEstimate();
+            telemetry.addLine(driveMethod.fieldState(myPose).toString());
 
             double drive = -gamepad1.left_stick_y;
             double strafe  = -gamepad1.left_stick_x;
             double rotate = gamepad1.right_stick_x;
 
+            //////////////GAMEPAD 1//////////////
             if(gamepad1.dpad_up) {
                 if(releasedDU1) {
                     increaseSpeed(0.05);
@@ -121,6 +140,28 @@ public class Mecanum_TeleOp_Final extends LinearOpMode {
             } else if (!releasedDD1){
                 releasedDD1 = true;
             }
+            if (gamepad1.left_bumper) {
+                if (releasedLB1 && Slide.getCurrentPosition() < initialHeight + 30){
+                    intakePower = 1.0;
+                    telemetry.addLine("INTAKE STARTS");
+                    releasedLB1 = false;
+                }
+            } else if (!releasedLB1){
+                intakePower = 0;
+                telemetry.addLine("INTAKE STOPS");
+                releasedLB1 = true;
+            }
+            if (gamepad1.right_bumper) {
+                if (releasedRB1 && Slide.getCurrentPosition() < initialHeight + 30){
+                    intakePower = -1.0;
+                    telemetry.addLine("INTAKE REVERSE STARTS");
+                    releasedRB1 = false;
+                }
+            } else if (!releasedRB1){
+                telemetry.addLine("INTAKE STOPS");
+                intakePower = 0;
+                releasedRB1 = true;
+            }
 
             if(gamepad1.a){
                 if(releasedA1) {
@@ -132,7 +173,7 @@ public class Mecanum_TeleOp_Final extends LinearOpMode {
 
             if(gamepad1.b){
                 if(releasedB1) {
-                    speed = 0.8;
+                    speed = 0.7;
                 }
             } else if(!releasedB1){
                 releasedB1 = true;
@@ -172,6 +213,18 @@ public class Mecanum_TeleOp_Final extends LinearOpMode {
                 releasedY2 = true;
             }
 
+            if(gamepad2.x){
+                if(releasedX2) {
+                    Rotate.setPosition(0.03);
+                    sleep(300);
+                    Slide.setTargetPosition(initialHeight);
+                    Slide.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+                    Slide.setPower(0.8);
+                }
+            } else if(!releasedX2){
+                releasedX2 = true;
+            }
+
             if (gamepad1.x) {
                 if (releasedX1){
                     if (toggleX1) {
@@ -195,6 +248,64 @@ public class Mecanum_TeleOp_Final extends LinearOpMode {
                 releasedX1 = true;
             }
 
+            //SHARED HUB
+            if(gamepad1.y) {
+                if (releasedY1) {
+                    if (PoseStorage.state != driveMethod.poseState.BLUE_SHARED_HUB && PoseStorage.state != driveMethod.poseState.RED_SHARED_HUB) {
+                        if (PoseStorage.state == driveMethod.poseState.BLUE || PoseStorage.state == driveMethod.poseState.BLUE_WAREHOUSE) {
+                            chassis.setPoseEstimate(fieldConstant.SHARED_BLUE_ENTER_POSE);
+                            Trajectory sharedTraj1 = chassis.trajectoryBuilder(chassis.getPoseEstimate(), true)
+                                    .lineTo(new Vector2d(64.75, 18))
+                                    .build();
+                            chassis.followTrajectory(sharedTraj1);
+                            Trajectory sharedTraj2 = chassis.trajectoryBuilder(sharedTraj1.end(), true)
+                                    .lineToLinearHeading(fieldConstant.SHARED_BLUE_END_POSE)
+                                    .build();
+                            chassis.followTrajectory(sharedTraj2);
+                        } else if (PoseStorage.state == driveMethod.poseState.RED || PoseStorage.state == driveMethod.poseState.RED_WAREHOUSE) {
+                            chassis.setPoseEstimate(fieldConstant.SHARED_RED_ENTER_POSE);
+                            Trajectory sharedTraj1 = chassis.trajectoryBuilder(chassis.getPoseEstimate(), true)
+                                    .lineTo(new Vector2d(64.75, 18))
+                                    .build();
+                            chassis.followTrajectory(sharedTraj1);
+                            Trajectory sharedTraj2 = chassis.trajectoryBuilder(sharedTraj1.end(), true)
+                                    .lineToLinearHeading(fieldConstant.SHARED_RED_END_POSE)
+                                    .build();
+                            chassis.followTrajectory(sharedTraj2);
+                        }
+                        speed = 0.3;
+                        chassis.updatePoseEstimate();
+                        PoseStorage.state = driveMethod.fieldState(chassis.getPoseEstimate());
+                    } else {
+                        if (PoseStorage.state == driveMethod.poseState.BLUE_SHARED_HUB) {
+                            Trajectory leaveTraj1 = chassis.trajectoryBuilder(chassis.getPoseEstimate())
+                                    .lineToLinearHeading(new Pose2d(64.75, chassis.getPoseEstimate().getY() + 1, toRadians(90)))
+                                    .build();
+                            chassis.followTrajectory(leaveTraj1);
+                            Trajectory leaveTraj2 = chassis.trajectoryBuilder(leaveTraj1.end())
+                                    .forward(24)
+                                    .build();
+                            chassis.followTrajectory(leaveTraj2);
+                        } else if (PoseStorage.state == driveMethod.poseState.RED_SHARED_HUB) {
+                            Trajectory leaveTraj1 = chassis.trajectoryBuilder(chassis.getPoseEstimate())
+                                    .lineToLinearHeading(new Pose2d(64.75, chassis.getPoseEstimate().getY() - 1, toRadians(-90)))
+                                    .build();
+                            chassis.followTrajectory(leaveTraj1);
+                            Trajectory leaveTraj2 = chassis.trajectoryBuilder(leaveTraj1.end())
+                                    .forward(24)
+                                    .build();
+                            chassis.followTrajectory(leaveTraj2);
+                        }
+                        speed = 0.7;
+                        chassis.updatePoseEstimate();
+                        PoseStorage.state = driveMethod.fieldState(chassis.getPoseEstimate());
+                    }
+                } else if (!releasedY1) {
+                    releasedY1 = true;
+                }
+            }
+
+            //////////////GAMEPAD 2//////////////
             if (gamepad2.dpad_up) {
                 if (releasedDU2){
                     Slide.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
@@ -220,8 +331,12 @@ public class Mecanum_TeleOp_Final extends LinearOpMode {
 
             if (gamepad2.dpad_left) {
                 if (releasedDL2){
-                    if(spinPower != 0) {
-                        spinPower -= 0.05;
+//                    if(spinPower != 0) {
+//                        spinPower -= 0.05;
+//                    }
+                    while(Rotate.getPosition() <= 1.0 && gamepad2.dpad_left) {
+                        Rotate.setPosition(Rotate.getPosition()+0.02);
+                        sleep(40);
                     }
                     releasedDL2 = false;
                 }
@@ -230,36 +345,17 @@ public class Mecanum_TeleOp_Final extends LinearOpMode {
             }
             if (gamepad2.dpad_right) {
                 if (releasedDR2){
-                    if(spinPower != 0) {
-                        spinPower += 0.05;
+//                    if(spinPower != 0) {
+//                        spinPower += 0.05;
+//                    }
+                    while(Rotate.getPosition() >= 0.03 && gamepad2.dpad_right) {
+                        Rotate.setPosition(Rotate.getPosition()-0.02);
+                        sleep(40);
                     }
                     releasedDR2 = false;
                 }
             } else if (!releasedDR2){
                 releasedDR2 = true;
-            }
-
-            if (gamepad1.left_bumper) {
-                if (releasedLB1 && Slide.getCurrentPosition() < initialHeight + 30){
-                    intakePower = 1.0;
-                    telemetry.addLine("INTAKE STARTS");
-                    releasedLB1 = false;
-                }
-            } else if (!releasedLB1){
-                intakePower = 0;
-                telemetry.addLine("INTAKE STOPS");
-                releasedLB1 = true;
-            }
-            if (gamepad1.right_bumper) {
-                if (releasedRB1 && Slide.getCurrentPosition() < initialHeight + 30){
-                    intakePower = -1.0;
-                    telemetry.addLine("INTAKE REVERSE STARTS");
-                    releasedRB1 = false;
-                }
-            } else if (!releasedRB1){
-                telemetry.addLine("INTAKE STOPS");
-                intakePower = 0;
-                releasedRB1 = true;
             }
 
             if (gamepad2.right_bumper) {
