+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/FreightFrenzy_2021/competition/delayed/Mecanum_Auto_BlueDuck_Delayed.java
@@ -1,10 +1,11 @@
-package org.firstinspires.ftc.teamcode.FreightFrenzy_2021.competition;
+package org.firstinspires.ftc.teamcode.FreightFrenzy_2021.competition.delayed;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
 import com.acmerobotics.roadrunner.trajectory.Trajectory;
 import com.qualcomm.hardware.bosch.BNO055IMU;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
@@ -28,6 +29,7 @@ import java.util.List;
 import static java.lang.Math.toRadians;
 
 @Autonomous(name = "BLUE DUCK DELAYED", group = "Competition")
+@Disabled
 public class Mecanum_Auto_BlueDuck_Delayed extends LinearOpMode {
 
     private DcMotor LF = null;
