+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/FreightFrenzy_2021/competition/delayed/Mecanum_Auto_RedDuck_Delayed.java
@@ -1,10 +1,11 @@
-package org.firstinspires.ftc.teamcode.FreightFrenzy_2021.competition;
+package org.firstinspires.ftc.teamcode.FreightFrenzy_2021.competition.delayed;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
 import com.acmerobotics.roadrunner.trajectory.Trajectory;
 import com.qualcomm.hardware.bosch.BNO055IMU;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
@@ -23,13 +24,13 @@ import org.firstinspires.ftc.robotcore.external.tfod.TFObjectDetector;
 import org.firstinspires.ftc.teamcode.roadrunner.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.robot_common.Robot4100Common;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import static java.lang.Math.toRadians;
 
-@Autonomous(name = "BLUE BARRIER DELAYED", group = "Competition")
-public class Mecanum_Auto_BlueBarrier_Delayed extends LinearOpMode {
+@Autonomous(name = "RED DUCK DELAYED", group = "Competition")
+@Disabled
+public class Mecanum_Auto_RedDuck_Delayed extends LinearOpMode {
 
     private DcMotor LF = null;
     private DcMotor RF = null;
@@ -40,7 +41,6 @@ public class Mecanum_Auto_BlueBarrier_Delayed extends LinearOpMode {
     private DcMotor Slide = null;
     private Servo Rotate = null;
     private Servo Push = null;
-    private ArrayList<Double[]> speedList = new ArrayList<Double[]>();
     private ElapsedTime runtime = new ElapsedTime();
 
     BNO055IMU imu;
@@ -57,9 +57,6 @@ public class Mecanum_Auto_BlueBarrier_Delayed extends LinearOpMode {
     private VuforiaLocalizer vuforia;
     private TFObjectDetector tfod;
 
-    double rotate = 0;
-    double speed = 0.5;
-    boolean reverse = false;
 
     @Override
     public void runOpMode() {
@@ -124,10 +121,10 @@ public class Mecanum_Auto_BlueBarrier_Delayed extends LinearOpMode {
 
         //Traj
         SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
-        Pose2d startPose = new Pose2d(6.5, 62.125, toRadians(90));
+        Pose2d startPose = new Pose2d(-41, -62.125, toRadians(-90));
         drive.setPoseEstimate(startPose);
         Trajectory myTrajectory1 = drive.trajectoryBuilder(startPose,true)
-                .splineToConstantHeading(new Vector2d(-11.875, 43), toRadians(-90))
+                .splineToConstantHeading(new Vector2d(-11.875, -43), toRadians(90))
                 .build();
 
         waitForStart();
@@ -139,7 +136,7 @@ public class Mecanum_Auto_BlueBarrier_Delayed extends LinearOpMode {
                 recogTime.reset();
             }
 
-            while (recogTime.milliseconds() <= 12000.0) {
+            while (recogTime.milliseconds() <= 10000.0) {
                 if (tfod != null) {
                     List<Recognition> updatedRecognitions = tfod.getUpdatedRecognitions();
                     if (updatedRecognitions != null) {
@@ -157,18 +154,18 @@ public class Mecanum_Auto_BlueBarrier_Delayed extends LinearOpMode {
                 }
             }
             if (center < 0) {
-                visionResult = "LEFT";
-            } else if (center < 321.85) {
-                visionResult = "MIDDLE";
-            } else {
                 visionResult = "RIGHT";
+            } else if (center < 420.725) {
+                visionResult = "LEFT";
+            } else {
+                visionResult = "MIDDLE";
             }
             telemetry.addLine(visionResult);
             telemetry.update();
 
             //MOTION TO PLATE
             drive.followTrajectory(myTrajectory1);
-            sleep(600);
+            sleep(500);
 
             //ROTATE
             Rotate.setPosition(1.0);
@@ -209,13 +206,22 @@ public class Mecanum_Auto_BlueBarrier_Delayed extends LinearOpMode {
 
             //BACK TO WALL
             Trajectory myTrajectory3 = drive.trajectoryBuilder(myTrajectory2.end())
-                    .splineToLinearHeading(new Pose2d(6.5, 65), toRadians(-180))
+                    .splineTo(new Vector2d(-63, -56), toRadians(-180))
                     .build();
             drive.followTrajectory(myTrajectory3);
-            sleep(300);
+            sleep(500);
 
+//            Pose2d wall = new Pose2d(-63.88, -56, Math.toRadians(90));
+//            drive.setPoseEstimate(wall);
+
+            //ROTATE DUCK
+            Spin.setPower(-0.5);
+            sleep(3500);
+            Spin.setPower(0.0);
+
+            //PARK
             Trajectory myTrajectory4 = drive.trajectoryBuilder(myTrajectory3.end())
-                    .forward(35)
+                    .strafeRight(20)
                     .build();
             drive.followTrajectory(myTrajectory4);
 
@@ -444,44 +450,6 @@ public class Mecanum_Auto_BlueBarrier_Delayed extends LinearOpMode {
         }
         stopMotion();
     }
-    void driveStrafeLeft(boolean isForward, double margin, double power, double timeInterval) {
-        ElapsedTime driveTime = new ElapsedTime(ElapsedTime.Resolution.MILLISECONDS);
-        final double currentAngle = aquireHeading();
-        int straightFactor = -1;
-        if (isForward) {
-            straightFactor = 1;
-        }
-        double targetAngle = currentAngle;
-        double LF_power;
-        double LB_power;
-        double RF_power;
-        double RB_power;
-        while (driveTime.milliseconds() < timeInterval) {
-
-            LF_power = straightFactor * power;
-            LB_power = straightFactor * power;
-            RF_power = straightFactor * power;
-            RB_power = straightFactor * power;
-
-            RF_power = Range.clip(RF_power, -1, 1);
-            RB_power = Range.clip(RB_power, -1, 1);
-            LF_power = Range.clip(LF_power, -1, 1);
-            LB_power = Range.clip(LB_power, -1, 1);
-            LF.setPower(0);
-            RF.setPower(RF_power);
-            LB.setPower(-LB_power);
-            RB.setPower(0);
-            telemetry.addData("RF_power", RF_power);
-            telemetry.addData("RB_power", RB_power);
-            telemetry.addData("LF_power", -LF_power);
-            telemetry.addData("LB_power", -LB_power);
-
-            telemetry.update();
-            displayEncoderValue();
-        }
-        stopMotion();
-    }
-
     //make a turn that based on the current heading in a certain direction and angle
     void rotateAtAngle(boolean isClockwise, double degree, double margin, double power) {
         int angleFactor = -1;
